+++ src/forze/application/usecases/document/__init__.py
@@ -8,12 +8,7 @@ backed by :class:`forze.application.contracts.document.DocumentPort` and
 from .create import CreateDocument
 from .delete import DeleteDocument, KillDocument, RestoreDocument, SoftDeleteArgs
 from .get import GetDocument
-from .list_ import (
-    RawListDocuments,
-    RawListDocumentsArgs,
-    TypedListDocuments,
-    TypedListDocumentsArgs,
-)
+from .list_ import RawListDocuments, TypedListDocuments
 from .update import UpdateArgs, UpdateDocument
 
 # ----------------------- #
@@ -29,6 +24,4 @@ __all__ = [
     "SoftDeleteArgs",
     "TypedListDocuments",
     "RawListDocuments",
-    "TypedListDocumentsArgs",
-    "RawListDocumentsArgs",
 ]