+++ src/forze/application/usecases/search/__init__.py
@@ -1,10 +1,5 @@
-from .read import RawSearch, RawSearchArgs, TypedSearch, TypedSearchArgs
+from .read import RawSearch, TypedSearch
 
 # ----------------------- #
 
-__all__ = [
-    "RawSearch",
-    "RawSearchArgs",
-    "TypedSearchArgs",
-    "TypedSearch",
-]
+__all__ = ["RawSearch", "TypedSearch"]